% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggHistogram.R
\name{ggHistogram}
\alias{ggHistogram}
\title{Ridged histogram for flow cytometry data.}
\usage{
ggHistogram(
  df,
  x,
  y,
  fill,
  facet = NULL,
  colors = NULL,
  facet_type = "wrap",
  facet_scales = "free",
  facet_space = "free",
  facet_ncol = NULL,
  facet_nrow = NULL
)
}
\arguments{
\item{df}{A data.frame of plot data.}

\item{x}{A string of the column name for x.}

\item{y}{A string of the column name for y.}

\item{fill}{A string of the column name for fill.}

\item{facet}{(Optional) A string of the column name for facet. Currently facetting with only onw variable is implemented.}

\item{colors}{(Optional) A vector of colors.}

\item{facet_type}{The facetting type. Can be either "wrap" or "grid".}

\item{facet_scales}{The facetting scales.}

\item{facet_space}{The spacing scheme in the facetting grid.}

\item{facet_ncol}{Number of columns.}

\item{facet_nrow}{Number of rows.}
}
\description{
The dashed line represents the median.
}
