% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggJitterPlot.R
\name{ggJitterPlot}
\alias{ggJitterPlot}
\title{Jitter plot with error bars.}
\usage{
ggJitterPlot(
  df,
  x,
  y,
  fill,
  shape = NULL,
  facet = NULL,
  colors = NULL,
  shapes = NULL,
  add.violin = "auto",
  minN.violin = 10,
  ylab = "Y",
  yscale = "continuous",
  yscale.expand = T,
  facet_type = "wrap",
  facet_scales = "free_y",
  facet_space = "fixed",
  facet_ncol = NULL,
  facet_nrow = NULL
)
}
\arguments{
\item{df}{A data.frame of plot data.}

\item{x}{A string of the column name for x.}

\item{y}{A string of the column name for y.}

\item{fill}{A string of the column name for fill.}

\item{shape}{(Optional) A string of the column name for shape.}

\item{facet}{(Optional) A string or a pair of strings of the column name(s) for facet.}

\item{colors}{(Optional) A vector of colors.}

\item{shapes}{(Optional) A vector of point shapes. E.g., 21:25}

\item{add.violin}{Logical. Whether a violin layer should be added. Default is "auto", in which case violin is generated when the maximal number of samples in a specific group >= minN.violin.}

\item{minN.violin}{An integer. The minimum number of data points required to add a violin layer.}

\item{ylab}{A string for y-axis label.}

\item{yscale}{The y-axis scale. Can be either "continuous" or "log10".}

\item{yscale.expand}{Logical. Whether the y-scale should be expanded down to zero.}

\item{facet_type}{The facetting type. Can be either "wrap" or "grid".}

\item{facet_scales}{The facetting scales.}

\item{facet_space}{The spacing scheme in the facetting grid.}

\item{facet_ncol}{Number of columns.}

\item{facet_nrow}{Number of rows.}
}
\description{
The error bar represents mean and SEM.
}
