% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtility.R
\docType{package}
\name{plotUtility}
\alias{plotUtility}
\alias{package-plotUtility}
\alias{plotUtility-package}
\title{plotUtility: utility functions for R graphics handling}
\description{
plotUtility: utility functions for R graphics handling
}
