% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_export.R
\name{utility_export}
\alias{utility_export}
\alias{savePDF}
\alias{savePPTX}
\title{Exporting plots as PDF or PowerPoint graphics}
\usage{
savePDF(
  graphicsObject = ggplot2::last_plot(),
  outputFileName = "plot.pdf",
  width = 8,
  height = 5,
  embedFont = F,
  gsPath = "C:/gs/gs9.16/bin/gswin32c.exe",
  gsOption =
    "-sFONTPATH=C:/Windows/Fonts -dCompressFonts=true -dSubsetFonts=false -dEmbedAllFonts=true"
)

savePPTX(
  graphicsObject = ggplot2::last_plot(),
  outputFileName = "plot.pptx",
  width = 8,
  height = 5
)
}
\arguments{
\item{graphicsObject}{The plot object to be saved. If missing, the most recently depicted plot will be retrieved using ggplot2::last_plot or recordPlot. Note: recordPlot is not compatible with exporting an editable vector graphic in PowerPoint.}

\item{outputFileName}{The name of the exported PDF file.}

\item{width}{A width.}

\item{height}{A height.}

\item{embedFont}{Whether the fonts need to be embedded using Ghostscript. Default is False.}

\item{gsPath}{A path to the Ghostscript exe file.}

\item{gsOption}{A character vector containing options to Ghostscript.}
}
\description{
Exporting plots as PDF or PowerPoint graphics
}
