% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_ggplot.R
\name{utility_ggplot_theme}
\alias{utility_ggplot_theme}
\alias{theme_Publication}
\title{A publication-ready ggplot theme}
\usage{
theme_Publication(base_size = 14)
}
\arguments{
\item{base_size}{A default font size.}
}
\description{
A publication-ready ggplot theme
}
