% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Plot.R
\name{Utility_Plot_ggplot-scales}
\alias{Utility_Plot_ggplot-scales}
\alias{scale_color_Publication}
\alias{scale_colour_Publication}
\alias{scale_fill_Publication}
\title{Publication-ready ggplot scales}
\usage{
scale_color_Publication(palette = F)

scale_colour_Publication(palette = F)

scale_fill_Publication(palette = F)
}
\arguments{
\item{palette}{Logical. If \code{FALSE}, a discrete scale for ggplot is returned. If \code{TRUE}, a color palette is returned.}
}
\description{
Publication-ready ggplot scales
}
