% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Plot.R
\name{Utility_Plot_ggplot-significance}
\alias{Utility_Plot_ggplot-significance}
\alias{format_pvalue}
\title{Statistical significance}
\usage{
format_pvalue(p)
}
\description{
Statistical significance
}
