% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Plot.R
\name{Utility_Plot_ggplot-theme}
\alias{Utility_Plot_ggplot-theme}
\alias{theme_Publication}
\title{A publication-ready ggplot theme}
\usage{
theme_Publication(base_size = 14)
}
\arguments{
\item{base_size}{The base font size.}
}
\description{
A publication-ready ggplot theme
}
