% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtility.R
\docType{package}
\name{plotUtility}
\alias{plotUtility}
\alias{package-plotUtility}
\title{plotUtility: utility functions for R graphics handling}
\description{
In-house utility functions for data visualization.
}
