% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Plot.R
\name{Utility_Plot}
\alias{Utility_Plot}
\alias{savePDF}
\alias{savePPTX}
\title{Plot utilities.}
\usage{
savePDF(
  ggObject = NULL,
  outputFileName = "plot.pdf",
  width = 8,
  height = 5,
  embedFont = F,
  gsPath = "C:/gs/gs9.16/bin/gswin32c.exe",
 
    gsOption = "-sFONTPATH=C:/Windows/Fonts -dCompressFonts=true -dSubsetFonts=false -dEmbedAllFonts=true"
)

savePPTX(ggObject = NULL, outputFileName = "plot.pptx", width = 8, height = 5)
}
\arguments{
\item{ggObject}{The ggplot object to be saved. If missing, the most recently depicted plot will be retrieved using ggplot2::last_plot.}

\item{outputFileName}{The name of the exported PDF file.}

\item{width}{A width.}

\item{height}{A height.}

\item{embedFont}{Whether the fonts need to be embedded using Ghostscript. Default is False.}

\item{gsPath}{A path to the Ghostscript exe file.}

\item{gsOption}{A character vector containing options to Ghostscript.}
}
\description{
Plot utilities.
}
