% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Plot.R
\name{Utility_Plot}
\alias{Utility_Plot}
\alias{savePDF}
\alias{savePPTX}
\title{Plot utilities.}
\usage{
savePDF(graphicObject = NULL, outputFileName = "plot.pdf", width = 8,
  height = 5, pointsize = 12, gsEmbetFont = T,
  gsPath = "C:/gs/gs9.16/bin/gswin32c.exe",
  gsOption = "-sFONTPATH=C:/Windows/Fonts -dCompressFonts=true -dSubsetFonts=false -dEmbedAllFonts=true")

savePPTX(graphicObject = NULL, outputFileName = "plot.pptx")
}
\arguments{
\item{graphicObject}{The plot object to be saved. If missing, the most recently depicted plot will be saved.}

\item{outputFileName}{The name of the exported PDF file.}

\item{width}{A width.}

\item{height}{A height.}

\item{pointsize}{A pointsize.}

\item{gsEmbetFont}{Logical. Whether fonts should be embedded using ghostscript.}

\item{gsPath}{A path to the Ghostscript exe file.}

\item{gsOption}{A character vector containing options to Ghostscript.}
}
\description{
Plot utilities.
}
